Enhance deep hedging scripts and model configurations
- Updated `01_train_deep_hedging.py` and `02_evaluate_deep_hedging.py` to ensure the project root is included in the Python path when run as scripts, improving usability.
- Added `drop_remainder=True` in data loaders for consistent batch sizes in graph mode.
- Enhanced error handling in `02_evaluate_deep_hedging.py` to provide clearer messages when a model is not found in the registry.
- Adjusted `default_model_config` to set `dropout_rate` to 0 for compatibility with graph mode.
- Improved test assertions in `test_policy_network.py` and `test_risk_measure.py` for better validation of model behavior and outputs.

+++ src/rade_ml/models/deep_hedging/layers/policy_network.py
@@ -79,7 +79,7 @@ class HedgingPolicy(tf.keras.layers.Layer):
         """Return zero-initialised hidden states for all RNN cells."""
         states = []
         for cell in self.gru_cells:
-            states.append(cell.get_initial_state(batch_size=batch_size, dtype=tf.float32))
+            states.append(cell.get_initial_state(batch_size=batch_size))
         return states
 
     def step(
